refactor(SelectedCartItemProvider): 서버 상태(cartItem)와 클라이언트 상태(selectedCartItem) 분리
