+++ src/contexts/carItem/CartItemProvider.tsx
@@ -1,8 +1,8 @@
 import { createContext } from "react";
-import { CartItem } from "../types/type";
-import { FREE_SHIPPING_MIN_AMOUNT, SHIPPING_FEE } from "../constants";
-import { LoadingStatus, useCartItems } from "../hooks/useCartItems";
-import { useSelected } from "../hooks/useSelected";
+import { CartItem } from "../../types/type";
+import { FREE_SHIPPING_MIN_AMOUNT, SHIPPING_FEE } from "../../constants";
+import { LoadingStatus, useCartItems } from "../../hooks/useCartItems";
+// import { useSelected } from "../hooks/useSelected";
 interface CartItemContext {
   cartItems: CartItem[];
   loadingStatus: LoadingStatus;
@@ -13,9 +13,9 @@ interface CartItemContext {
   orderPrice: number;
   shippingFee: number;
   totalPrice: number;
-  selectedItemIds: Set<number>;
-  toggleSelectedItemId: (id: number) => void;
-  replaceSelectedItemIds: (ids: number[]) => void;
+  // selectedItemIds: Set<number>;
+  // toggleSelectedItemId: (id: number) => void;
+  // replaceSelectedItemIds: (ids: number[]) => void;
   handleLoadingStatus: (status: LoadingStatus) => void;
 }
 
@@ -35,15 +35,17 @@ export const CartItemProvider = ({ children }: CartItemProviderProps) => {
     errorMessage,
     handleLoadingStatus,
   } = useCartItems();
-  const { selectedItemIds, toggleSelectedItemId, replaceSelectedItemIds } =
-    useSelected({ enableStorage: true, storageKey: "selectedCartItemIds" });
+  // const { selectedItemIds, toggleSelectedItemId, replaceSelectedItemIds } =
+  //   useSelected({ enableStorage: true, storageKey: "selectedCartItemIds" });
 
-  const orderPrice = cartItems.reduce((acc, cartItem) => {
-    if (selectedItemIds.has(cartItem.id)) {
-      return acc + cartItem.product.price * cartItem.quantity;
-    }
-    return acc;
-  }, 0);
+  // const orderPrice = cartItems.reduce((acc, cartItem) => {
+  //   if (selectedItemIds.has(cartItem.id)) {
+  //     return acc + cartItem.product.price * cartItem.quantity;
+  //   }
+  //   return acc;
+  // }, 0);
+
+  const orderPrice = 0;
 
   const shippingFee = orderPrice >= FREE_SHIPPING_MIN_AMOUNT ? 0 : SHIPPING_FEE;
 
@@ -61,9 +63,9 @@ export const CartItemProvider = ({ children }: CartItemProviderProps) => {
         orderPrice,
         shippingFee,
         totalPrice,
-        selectedItemIds,
-        toggleSelectedItemId,
-        replaceSelectedItemIds,
+        // selectedItemIds,
+        // toggleSelectedItemId,
+        // replaceSelectedItemIds,
         handleLoadingStatus,
       }}
     >
